# swagger/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API do BoardTasks",
    "description": "Documentação da API do BoardTasks."
  },
  "servers": [
    {
      "url": "http://localhost:4000/",
      "description": "Servidor localhost."
    },
    {
      "url": "https://dnc-board-tasks-back.vercel.app/",
      "description": "Servidor de produção."
    }
  ],
  "paths": {
    "/usuario/criar": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "senha": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuario/logar": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "senha": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tarefa/criar": {
      "post": {
        "tags": [
          "Tarefa"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "posicao": {
                    "example": "any"
                  },
                  "titulo": {
                    "example": "any"
                  },
                  "descricao": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "dataEntrega": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tarefa/editar/{id}": {
      "put": {
        "tags": [
          "Tarefa"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "posicao": {
                    "example": "any"
                  },
                  "titulo": {
                    "example": "any"
                  },
                  "descricao": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "dataEntrega": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tarefa/obter/usuario": {
      "get": {
        "tags": [
          "Tarefa"
        ],
        "description": "Endpoint para obter todas tarefas do usuario logado.",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tarefa/deletar/{id}": {
      "delete": {
        "tags": [
          "Tarefa"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Usuario"
          },
          "required": {
            "type": "array",
            "example": [
              "nome",
              "email",
              "senha"
            ],
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "properties": {
              "nome": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "email": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "senha": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "createdAt": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "updatedAt": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      },
      "Tarefa": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Tarefa"
          },
          "required": {
            "type": "array",
            "example": [
              "posicao",
              "titulo",
              "status",
              "usuarioCriador"
            ],
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "properties": {
              "posicao": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "titulo": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "descricao": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "status": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "dataEntrega": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "usuarioCriador": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "createdAt": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              },
              "updatedAt": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}